initial refactor of profile models
